code cleanup and ui edits
