+++ client/src/components/Profile/ScheduleHistory/ScheduleHistory.jsx
@@ -1,14 +1,15 @@
-import { Box, Pagination } from '@mui/material';
-import React, { useEffect, useState } from 'react';
+import { Grid, Pagination } from '@mui/material';
+import React, { useEffect, useState, useReducer } from 'react';
 
 import ScheduleItem from './ScheduleItem';
-import workoutAndMealScheduleService from '../../../services/WorkoutAndMealScheduleService';
+import userProfileService from '../../../services/UserProfileService';
 
 export default function ScheduleHistory() {
   const [history, setHistory] = useState(null);
+  const [token, forceUpdate] = useReducer((x) => x + 1, 0);
 
   const getSchedules = () => {
-    workoutAndMealScheduleService.getPaginatedWorkoutAndMealSchedule(1)
+    userProfileService.getPaginatedScheduleHistory(1)
       .then((res) => {
         setHistory(res.response.data.pagination);
       });
@@ -16,10 +17,11 @@ export default function ScheduleHistory() {
 
   // eslint-disable-next-line no-unused-vars
   const handlePageChange = async (_, page) => {
-    workoutAndMealScheduleService.getPaginatedWorkoutAndMealSchedule(page)
+    userProfileService.getPaginatedScheduleHistory(page)
       .then((res) => {
         setHistory(res.response.data.pagination);
       });
+    forceUpdate();
   };
 
   useEffect(() => {
@@ -27,16 +29,16 @@ export default function ScheduleHistory() {
   }, []);
 
   return (
-    <Box sx={{
-      width: '100%', height: '100%', display: 'flex', flexDirection: 'column', overflowY: 'auto',
-    }}
+    <Grid
+      container
+      spacing={2}
     >
       {
         history
           // eslint-disable-next-line react/no-array-index-key
-          ? history.items.map((i, idx) => <ScheduleItem key={`schedule-${idx}`} details={i} />) : ''
+          ? history.schedules.map((i, idx) => <ScheduleItem key={`schedule-${idx}`} details={i} token={token} />) : ''
       }
       <Pagination count={history ? history.max : 1} page={history ? history.page : 1} onChange={handlePageChange} sx={{ marginTop: '20px' }} />
-    </Box>
+    </Grid>
   );
 }
