+++ client/src/services/WorkoutAndMealScheduleService.js
@@ -19,16 +19,6 @@ const getWorkoutAndMealSchedule = async () => {
   };
 };
 
-const getPaginatedWorkoutAndMealSchedule = async (page) => {
-  const response = await axios.post(`${process.env.REACT_APP_SERVER_API_BASE}/mock/workoutMealSchedule/all`, { page }, {
-    headers: createBearerTokenHeader(),
-  });
-
-  return {
-    response,
-  };
-};
-
 const createWorkoutAndMealSchedule = async () => {
   const apiUrl = `${process.env.REACT_APP_SERVER_API_BASE}/userProfile/generate/`;
 
@@ -42,7 +32,6 @@ const createWorkoutAndMealSchedule = async () => {
 const workoutAndMealScheduleService = {
   createWorkoutAndMealSchedule,
   getWorkoutAndMealSchedule,
-  getPaginatedWorkoutAndMealSchedule,
 };
 
 export default workoutAndMealScheduleService;
