+++ server/src/routes/mock.js
@@ -5,7 +5,7 @@ const router = express.Router();
 const { protect } = require('../middleware/authMiddleware');
 const { MOCKcreateMealScheduleForUser, MOCKupdateMealScheduleForUser } = require('../controllers/workoutSchedule/mockMealScheduleController');
 const { MOCKupdateUserWorkoutScheduleByUserID, MOCKcreateWorkoutSchedule } = require('../controllers/workoutSchedule/mockWorkoutScheduleController');
-const { MOCKcreateWorkoutAndMealSchedule, MOCKgetPaginatedWorkoutAndMealSchedule } = require('../controllers/workoutSchedule/mockWorkoutAndMealSchedule');
+const { MOCKcreateWorkoutAndMealSchedule } = require('../controllers/workoutSchedule/mockWorkoutAndMealSchedule');
 
 /**
  * @desc use mock data to create a user schedule (NO OPENAI CALL)
@@ -164,7 +164,5 @@ router.put('/workoutSchedule', protect, MOCKupdateUserWorkoutScheduleByUserID);
  *
  */
 router.post('/workoutMealSchedule', protect, MOCKcreateWorkoutAndMealSchedule);
-
-router.post('/workoutMealSchedule/all', MOCKgetPaginatedWorkoutAndMealSchedule);
 
 module.exports = router;
